% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scheme_vector.R
\name{color_scheme_vector}
\alias{color_scheme_vector}
\title{Generate color scheme as named vector}
\usage{
color_scheme_vector(color_list, color_by, long_count_data)
}
\arguments{
\item{color_list}{A list of colors from unique_colors().}

\item{color_by}{Attribute to color by. A column from long_count_data, inputted as a string.}

\item{long_count_data}{Long form data, e.g. an output from get_nmds_scores().}
}
\description{
This function creates a color scheme formatted as a named vector
}
\examples{
color_scheme_vector
}
\keyword{color}
\keyword{named}
\keyword{scheme}
\keyword{vector}
