% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edges.R
\name{generate_edges}
\alias{generate_edges}
\title{Generate edges for network analysis}
\usage{
generate_edges(abundance_matrix, nodes, dissimilarity, cutoff)
}
\arguments{
\item{abundance_matrix}{An OTU or ASV abundance table.}

\item{nodes}{A nodes() object.}

\item{dissimilarity}{A defined dissimilarity metric; c("spearman", "kendall", "pearson") accepted.}

\item{cutoff}{A defined asymptotic p-value cutoff for each edge. Rows with values higher than this cutoff are discarded.}
}
\description{
This function creates a list of network edges from an OTU or ASV abundance table. See nodes()
}
\examples{
generate_edges()
}
\keyword{edges}
\keyword{network}
