% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{generate_nodes}
\alias{generate_nodes}
\title{Generate nodes for network analysis}
\usage{
generate_nodes(asv_table)
}
\arguments{
\item{asv_table}{An OTU or ASV table.}
}
\description{
This function generates network nodes for visNetwork from an OTU or ASV table.
}
\examples{
generate_nodes()
}
\keyword{network}
\keyword{nodes}
