% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmds_scores.R
\name{get_nmds_scores}
\alias{get_nmds_scores}
\title{Get NMDS scores}
\usage{
get_nmds_scores(metaMDS_output, metadata, joining_variable)
}
\arguments{
\item{metaMDS_output}{A metaMDS object.}

\item{metadata}{Metadata file.}

\item{joining_variable}{A defined column by which scores and metadata will merge. Should be consistent with grouping_variable argument from calc_taxa_mean().}
}
\description{
This function extracts NMDS scores and merges associated metadata file for ease of ggplot2-based plotting.
}
\examples{
get_nmds_scores()
}
\keyword{color}
\keyword{scheme}
\keyword{unique}
