% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmds_vectors.R
\name{get_nmds_vectors}
\alias{get_nmds_vectors}
\title{Extract environmental vectors for NMDS plotting}
\usage{
get_nmds_vectors(env_output, top, pvalue)
}
\arguments{
\item{env_output}{Output object from envfit(),}

\item{top}{Top n significant vectors to display.}

\item{pvalue}{Defined p-value cutoff.}
}
\description{
This function takes a vegan::envfit() output and extracts the top "n" significant vectors as a dataframe.
}
\examples{
get_nmds_vectors
}
\keyword{nmds}
\keyword{vectors}
