% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_asv_table.R
\name{import_asv_table}
\alias{import_asv_table}
\title{Import qiime2 artifact as ASV table dataframe}
\usage{
import_asv_table(file_path)
}
\arguments{
\item{file_path}{Path to a qiime2 .qza file.}
}
\description{
This function uses qiime2R to create an OTU or ASV table as a dataframe.
}
\examples{
import_asv_table()
}
\keyword{asv}
\keyword{import}
\keyword{table}
