% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Create a matrix from long form data}
\usage{
make_matrix(long_data, count_column_name, grouping_name)
}
\arguments{
\item{long_data}{Long-form data to be converted into matrix.}

\item{count_column_name}{Name of column that contains count data. May be relative or absolute count data.}

\item{grouping_name}{Name of column to be used as rownames. Should be consistent with grouping_variable used in calc_taxa_mean.}
}
\description{
This function creates an OTU or ASV matrix from long form data (e.g., output from calc_taxa_mean). Output is usable with matrix-based distance calculations.
}
\examples{
make_matrix()
}
\keyword{matrix}
