% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_taxa.R
\name{rare_taxa}
\alias{rare_taxa}
\title{Return rare taxa}
\usage{
rare_taxa(asv_table, abun.cutoff)
}
\arguments{
\item{asv_table}{An OTU or ASV abundance table.}

\item{abun.cutoff}{A defined relative relative abundance cutoff. Only taxa with a relative abundance less than this value (a value from 0 to 1) are returned.}
}
\description{
This function filters an ASV or OTU table by a defined relative abundance cutoff such that only counts less than the cutoff are returned.
}
\examples{
rare_taxa
}
\keyword{abundance}
\keyword{cutoff}
\keyword{relative}
