% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_taxon_id.R
\name{split_taxonID}
\alias{split_taxonID}
\title{Split taxonomic ID}
\usage{
split_taxonID(asv_table)
}
\arguments{
\item{asv_table}{An OTU or ASV abundance table.}
}
\description{
This function takes an ASV table and returns a longform dataframe of counts with separate ID columns for each level of taxonomic classification.
}
\examples{
split_taxonID
}
\keyword{split}
\keyword{taxonID}
