% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_colors.R
\name{unique_colors}
\alias{unique_colors}
\title{Generate list of unique colors}
\usage{
unique_colors(df, color_by, brewer.pal_name)
}
\arguments{
\item{df}{An object of class dataframe.}

\item{color_by}{Attribute to color by.}

\item{brewer.pal_name}{RColorBrewer palette to use.}
}
\description{
This function generates a color palette. Uses RColorBrewer.
}
\examples{
unique_colors()
}
\keyword{color}
\keyword{scheme}
\keyword{unique}
